refactor: 금액에 관련해서 구매로직에 해당하는 부분을 BigInteger로 변경한다

+++ src/main/java/lottogame/controller/LottoController.java
@@ -1,5 +1,6 @@
 package lottogame.controller;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -29,7 +30,7 @@ public class LottoController {
     }
 
     private List<LottoTicketDto> purchaseLottoTickets() {
-        long money = inputer.inputMoney();
+        BigInteger money = inputer.inputMoney();
         int passiveLottoTicketCount = inputer.inputPassiveLottoTicketCount();
         List<LottoTicketDto> selectedLottoTickets = toLottoTicketDtos(
             inputer.inputPassiveLottoTickets(passiveLottoTicketCount));
